+++ tests/settings.py
@@ -128,3 +128,8 @@ FLOW_PROCESSES_FINDERS = (
 )
 
 FLOW_PROCESSES_DIRS = (os.path.join(PROJECT_ROOT, '../resolwe_bio/tests/'),)
+
+# Do not skip tests that fail on Docker executor if this is set via environment
+# variable
+if os.environ.get('RESOLWEBIO_TESTS_SKIP_DOCKER_FAILURES', '').lower() in ["no", "false"]:
+    TESTS_SKIP_DOCKER_FAILURES = False